+++ metricsbp/BUILD.bazel
@@ -7,9 +7,9 @@ go_library(
         "config.go",
         "doc.go",
         "nil_check.go",
+        "runtime_stats.go",
         "sampled.go",
         "statsd.go",
-        "sys_stats.go",
         "tags.go",
         "timer.go",
     ],
